[WinUI] Fix resizing CV rows when content size changed
When measuring the ItemContentControl we might not get the right value if the size of its content just changed. In order to ensure we get the up to date value, we first measure the content (to update DesiredSize), and then we measure the ItemContentControl.

Fixes https://github.com/dotnet/maui/issues/18078